# swaggerLog.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ecommerce API",
    "version": "1.0.0",
    "description": "The Demo ecommerce api made for user, cart, and order transactions for an art products store."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Main server"
    }
  ],
  "paths": {
...
    "/api/register": {
      "summary": "Input user data to register.",
      "description": "This path is for creating an account to add items to user cart for further purchases.",
      "post": {
        "summary": "Creates an account for a new user.",
        "description": "Creates an account for a new user.",
        "operationId": "post_register",
        "tags": ["Authentication"],
        "requestBody": {
          "description": "Object for creating new user. Password must be at least 6 characters long. Email must match to email syntax.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              },
              "example": {
                "email": "test_for_test@example.com",
                "password": "test123456",
                "first_name": "Test",
                "last_name": "Testtest",
                "address": "Test, 00/11",
                "postcode": "010101",
                "city": "Test",
                "country": "Test"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "text/plain": {
                "example": "User created"
              }
            }
          },
          "400": {
            "description": "Invalid entries in request body",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Email is already in use"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "summary": "Login to your account.",
      "description": "This path is for loging in to your account to access account functions.",
      "post": {
        "summary": "Login to your account.",
        "description": "Enter your email and password to login with Passport and access your user info, cart, and orders. Passport starts a session and keeps you logged in until you send a GET request to /api/logout.",
        "operationId": "post_login",
        "tags": ["Authentication"],
        "requestBody": {
          "description": "Object for logging into account. Password must be at least 6 characters long. Email must match to email syntax.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "example": {
                "email": "test_for_test@example.com",
                "password": "test123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User is logged in.",
            "content": {
              "text/plain": {
                "example": "Token. Message: Test is logged in, customer"
              }
            }
          },
          "400": {
            "description": "Invalid entries in request body.",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Email/Password is not found"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/carts": {
      "description": "Access users carts",
      "get": {
        "summary": "Get user cart information.",
        "description": "This path fetches user cart information; also fetches multiple carts if user has multiple carts. Uses user id from passport Loggin.",
        "operationId": "get_carts",
        "tags": ["Cart"],
        "responses": {
          "200": {
            "description": "Returns array of user's cart object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Carts"
                },
                "example": [
                  {
                    "id": 4,
                    "user_id": 22,
                    "created_at": "2022-05-18T13:54:53.660Z",
                    "modified": "2022-06-04T10:25:39.235"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "No any carts under your user id",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "No any carts found"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Creates a new cart.",
        "description": "Creates a new cart using the user id provided by your Passport loggin.",
        "operationId": "post_carts",
        "tags": ["Cart"],
        "responses": {
          "201": {
            "description": "Cart is created",
            "content": {
              "text/plain": {
                "example": "Cart is created"
              }
            }
          }
        }
      }
    },
    "/api/carts/{cartId}": {
      "description": "Preform CRUD operations on a cart based on cart id from parameters and user id provided by your Passport loggin.",
      "get": {
        "summary": "Get cart by cart id.",
        "description": "This path fetches user cart information based on cart id and the user id provided by your Passport loggin.",
        "operationId": "get_carts_id",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Cart id.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns cart object matching to provided cart id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                },
                "example": {
                  "id": 3,
                  "user_id": 3,
                  "created": "2022-05-18",
                  "modified": null
                }
              }
            }
          },
          "400": {
            "description": "Invalid cart id.",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Invalid cart id"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete cart by cart id.",
        "description": "This path delete cart based on cart id from the parameters and user id provided by your Passport loggin.",
        "operationId": "delete_carts_id",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Cart id.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content - cart is deleted"
          }
        }
      }
    },
    "/api/carts/{cartId}/checkout": {
      "get": {
        "summary": "Submit cart for checkout",
        "description": "Creates a new order using all products in the selected cart (parameters) and the user id provided by your Passport loggin.",
        "operationId": "get_carts_checkout",
        "tags": ["Cart/Checkout"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Cart id.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns id of the new order that was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkout"
                },
                "example": {
                  "order_id": 4
                }
              }
            }
          },
          "400": {
            "description": "No items in cart",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Cart empty. No payment processed"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/carts/{cartId}/items": {
      "get": {
        "summary": "Gets all items in cart",
        "description": "This path fetches all items in the selected cart (cart id from parameter and user id provided by your Passport loggin).",
        "operationId": "get_cart_items",
        "tags": ["Cart/Items"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Cart id.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All product objects in cart returned in an array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductAllCart"
                },
                "example": [
                  {
                    "id": 1,
                    "name": "Abstract cityscape",
                    "author": "Andrej Ostapchuk",
                    "description": "The picture with a volume and a depth of space.",
                    "category": "Abstract",
                    "status": "Active",
                    "quantity": 1
                  },
                  {
                    "id": 15,
                    "name": "Dragon",
                    "author": "Netsuke",
                    "description": "Dragon. Ivory, Meiji-Period (1868 - 1912)",
                    "category": "Netsuke",
                    "status": "Active",
                    "price": 76499,
                    "quantity": 1
                  },
                  "..."
                ]
              }
            }
          },
          "400": {
            "description": "No items in cart",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Cart Empty"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add item to cart",
        "description": "Add item to the selected cart (cart id from parameter and user id provided by your Passport loggin).",
        "operationId": "post_cart_items",
        "tags": ["Cart/Items"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Cart id.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Submit product id and quantity in request body.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAdd"
              },
              "example": {
                "product_id": 1,
                "quantity": 2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added to cart",
            "content": {
              "text/plain": {
                "example": "Product added"
              }
            }
          },
          "400": {
            "description": "Invalid product id",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Invalid product Id"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "summary": "Get all orders.",
        "description": "This path fetches all orders for your account. User id for fetching accounts is provided by your Passport loggin.",
        "operationId": "get_orders",
        "tags": ["Orders"],
        "responses": {
          "200": {
            "description": "All order objects sent in an array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersAll"
                },
                "example": [
                  {
                    "id": 3,
                    "user_id": 22,
                    "status": "PENDING",
                    "created_at": "2022-05-18T15:41:11.354Z",
                    "modified": "2022-05-24T18:15:40.269Z",
                    "total_price": "$712,500.00"
                  },
                  {
                    "id": 4,
                    "user_id": 22,
                    "status": "PENDING",
                    "created_at": "2022-05-21T11:46:44.626Z",
                    "modified": "2022-05-24T18:16:20.614Z",
                    "total_price": "$189,990.00"
                  },
                  {
                    "id": 5,
                    "user_id": 22,
                    "status": "PENDING",
                    "created_at": "2022-05-21T12:01:18.897Z",
                    "modified": "2022-05-24T18:16:42.857Z",
                    "total_price": "$25,056.00"
                  },
                  "..."
                ]
              }
            }
          },
          "400": {
            "description": "No orders found.",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "No orders found"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "summary": "Get order by order id.",
        "description": "This path fetches an order from your account (user id provided by your Passport loggin) that matches the order id provided (parameters).\nProducts in order are attached in an array to the order object.",
        "operationId": "get_orders_id",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "Order id.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns order object with id matching the parameters. Products in order are attatched in array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {               
                    "id": 3,
                    "user_id": 22,
                    "status": "PENDING",
                    "created_at": "2022-05-18T15:41:11.354Z",
                    "modified": "2022-05-24T18:15:40.269Z",
                    "total_price": "$712,500.00",
                    "products": [
                      {
                        "order_items": "(3,1,15,\"$712,500.00\")",
                        "product_id": 1,
                        "quantity": 15
                      }
                    ]
                  }
                  
              }
            }
          },
          "400": {
            "description": "Order with that id is not found",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Order is not found"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Registration": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "Carts": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "user_id": {
              "type": "integer"
            },
            "created": {
              "type": "string"
            },
            "modified": {
              "type": "string"
            }
          }
        }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        }
      },
      "Checkout": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer"
          }
        }
      },
      "ProductAllCart": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "author": {
              "type": "string"
            },
            "price": {
              "type": "string"
            },
            "category": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "quantity": {
              "type": "integer"
            }
          }
        }
      },
      "ProductAdd": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          }
        }
      },
...
      "OrdersAll": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "user_id": {
              "type": "integer"
            },
            "order_price": {
              "type": "string"
            },
            "status": {
              "type": "string"
            },
            "created": {
              "type": "string"
            },
            "modified": {
              "type": "string"
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "order_price": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          },

          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "author": {
                  "type": "string"
                },
                "price": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    }
  }
}